move circom proof code into `primitives` and update `vanchor/src/tests_circom.rs` accordingly. Standalone node broke for some reason
