+++ primitives/Cargo.toml
@@ -23,14 +23,13 @@ ark-bls12-381 = { version = "^0.3.0", default-features = false, features = ["cur
 ark-bn254 = { version = "^0.3.0", default-features = false, features = ["curve"], optional = true }
 ark-ec = { version = "^0.3.0", default-features = false }
 ark-ff = { version = "^0.3.0", default-features = false }
-ark-groth16 = { version = "^0.3.0", default-features = false }
+ark-groth16 = { git = "https://github.com/arkworks-rs/groth16", rev = "765817f", default-features = false, features = ["parallel"] }
 ark-relations = { version = "^0.3.0", default-features = false }
 ark-serialize = { version = "^0.3.0", default-features = false, features = ["derive"] }
 ark-std = { version = "^0.3.0", default-features = false }
 # ark-circom = { git = "https://github.com/gakonst/ark-circom.git", default-features = false }
 
 byteorder = { version = "1", default-features = false }
-num-traits = { version = "0.2.0", default-features = false }
 blake2 = { version = "0.9", default-features = false }
 digest = { version = "0.9", default-features = false }
 
@@ -41,8 +40,18 @@ ethabi = { version = "15.0.0", default-features = false }
 
 webb-proposals = { git = "https://github.com/webb-tools/webb-rs", default-features = false, features = ["scale", "substrate", "evm"] }
 
+ark-circom = { git = "https://github.com/vacp2p/ark-circom", branch = "wasm", default-features = false, features = ["circom-2"] }
+num-bigint = { version = "0.4", default-features = false, features = ["rand"] }
+once_cell = "1.14.0"
+cfg-if = "1.0"
+wasmer = { version = "2.3.0", default-features = false, features = ["sys-default"]}
+num-traits = "0.2.11"
+thiserror = "1.0.0"
+color-eyre = "0.6.1"
+serde_json = "1.0.48"
+
 [features]
-default = ["std", "hashing", "verifying", "field_ops"]
+default = ["std", "hashing", "verifying", "field_ops", "wasmer/sys-default"]
 field_ops = [
   "ark-bls12-381",
   "ark-bn254",
@@ -73,3 +82,4 @@ verifying = [
   "ark-bls12-381",
   "ark-bn254",
 ]
+wasm = ["wasmer/js", "wasmer/std"]
